Ensure that member is owner when deleting house group

+++ API/src/SharedHome.Application/HouseGroups/Commands/DeleteHouseGroup/DeleteHouseGroupHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using SharedHome.Application.HouseGroups.Extensions;
+using SharedHome.Domain.HouseGroups.Exceptions;
 using SharedHome.Domain.HouseGroups.Repositories;
 using SharedHome.Domain.Invitations.Repositories;
 using SharedHome.Shared.Abstractions.Commands;
@@ -19,6 +20,11 @@ namespace SharedHome.Application.HouseGroups.Commands.DeleteHouseGroup
         public async Task<Response<Unit>> Handle(DeleteHouseGroupCommand request, CancellationToken cancellationToken)
         {
             var houseGroup = await _houseGroupRepository.GetOrThrowAsync(request.HouseGroupId, request.PersonId!);
+
+            if (!houseGroup.IsOwner(request.PersonId!))
+            {
+                throw new HouseGroupMemberIsNotOwnerException(request.PersonId);
+            }
 
             await _houseGroupRepository.DeleteAsync(houseGroup);
